+++ CHANGELOG.md
@@ -7,3 +7,9 @@ From v1.4.0 forwards, all notable changes to this project will be documented in 
 # 1.5.0
 - Allow more flexibly specifying command line args
 - Minor fix for logging
+
+# 1.6.0
+- Automatically get scaling parameter of monitors. Credit to [SetDpi](https://github.com/imniko/SetDPI)
+- Improve position calculation
+- Add DebugDisplay config
+- Allow hyper in config key at command line